feat: set up Claude/Anthropic models in Shellbot
`:Shellbot` replaces `:ChatGPT` as the default command now, but I set up
other commands to facilitate switching:

- `:ChatGPT` (gpt-4o)
- `:ChatGPTX` (o1-mini: slower, but supposedly better; too verbose IMO)
- `:Claude` (Sonnet)
- `:Opus` (older, slower model)

Likewise, shell aliases:

- `chatgpt` (same as before)
- `chatgptx` (same as before)
- `claude` (new)
- `opus` (new)

But beware: there is an interplay between these things. The aliases work
by _unsetting_ the environment variable for the corresponding API key
and then launching Neovim. That means that if you launch Neovim with,
say, `claude`, and then try to invoke `:ChatGPT`, it won't actually use
the OpenAI model. This is because the underlying shellbot call will look
in the environment and _not_ find the expected variable. I'm not too
worried about this edge case.

* aspects/nvim/files/.config/nvim/pack/bundle/opt/shellbot f848d28...eb6d184 (3):
  > docs: document new environment variables
  > chore: default Anthropic to "claude-3-5-sonnet-20241022"
  > refactor!: rename `:ChatGPT` to `:Shellbot`