refactor(nvim): set up treesitter textobjects
These are language dependent, so in some sense this is a bit "meh". For
example, not all languages define `@inner.comment`. So that means that
`vic` works in YAML, but not in TS, for instance.

Oh well. That's probably okay because I'm mostly using Vim's built in
motions anyway:

- https://neovim.io/doc/user/various.html#commenting

eg. `gcgc` to uncomment a block, `gc{motion}` to comment a region, `dgc`
to delete a comment, etc.

On the flip side, it does also mean that I get to delete some truly
ancient plug-ins.

Submodule changes - 1 added and 2 removed:

* aspects/nvim/files/.config/nvim/pack/bundle/opt/nvim-treesitter-textobjects 0000000...ad8f0a4 (519):
  > fix(julia): adapt queries to parser changes

* aspects/nvim/files/.config/nvim/pack/bundle/opt/vim-textobj-comment 58ae457...0000000:

* aspects/nvim/files/.config/nvim/pack/bundle/opt/vim-textobj-rubyblock 2b882e2...0000000:

* aspects/nvim/files/.config/nvim/pack/bundle/opt/vim-textobj-user 41a675d...0000000: