refactor(nvim): use treesitter for Markdown folding
In general, I like the indent-based folding better everywhere else
because:

- it is fast (because it's simple)
- it is robust (because it's simple)
- it works (because of source-code formatters, the indentation of
  everything I'm editing tends to be in good shape)

But Markdown was always an exception. It barely uses indentation, so the
heading-based folding provided by the vim-markdown plugin was better. We
can achieve something very similar though with treesitter: this will
fold sections, fenced code blocks etc. It appears to work quite well in
practice, and it's one less submodule to have to manage.

* aspects/nvim/files/.config/nvim/pack/bundle/opt/vim-markdown 8f6cb3a...0000000: