refactor(nvim): dumb-down window blur and focus
My first inclination was to update this to work with treesitter, but
there are a number of unpleasant things that would have to be dealt
with.

- We'd have to detect which buffers have treesitter active (and call
  `vim.treesitter.stop()` and which ones are eligible to have it
  activated (and call `vim.treesitter.start()`).
- If we could reliably do that (ideally without hacks), then we'd still
  have to cope with the fact that some buffers are going to use
  treesitter and others are going to use old-school syntax highlighting.
  It seems all too easy to wind up with both activated accidentally.
- I found that even `ownsyntax on`, `syntax off`, and
  `vim.treesitter.stop()` don't actually strip all the color out of a
  buffer. Something, somewhere is still matching the text and
  highlighting it.

Now, I could sit down and try to figure all this out, or I could just
resign myself to living with a bit more color in blurred buffers, and
be thankful that I've removed something brittle, fragile, and complex.
So I'm going to do the latter.