feat(nvim): enable "indent" in treesitter config
I initially didn't turn this on because I didn't have a handy example of
something that it handled well (beyond what Neovim seems to be doing
because `'autoindent'` is set).

I found at least one case that this handles better when enabled:

    const foo = (
        config: NonNullable<A>,
        value: number,
    ) => {
        const {
            width = DEFAULT_WIDTH,
            length = DEFAULT_LENGTH,
            hidden = false,
        } = config;
        // ...
    }

Probably not a minimal test case, but it's the one that I have at hand.
If you're on the comment line and open a line above (or if you're on the
line above and you try to open a line below), it will appear indented by
one extra shiftwidth.

There are a number of other indent-related things that don't seem to
work as well under treesitter as they did under the legacy plugins, but
I don't know if there's anything I can really do about that at this
point.