fix(nvim): actually calculate `position`
I committed a placeholder value last night because I was in a hurry and
had to rush out. Looping back now to actually implement the logic to
calculate the position.

I think it's probably correct, but in my testing it doesn't handle
removed fold markers very well. The state of the cache appears to be
correctly updated, but you still need to hit `zx` to get Neovim to
recompute things. Seems that this is somewhat expected, because Neovim
doesn't know when to call `foldexpr()` for all the lines that might have
been affected by a change:

    zx    Update folds: Undo manually opened and closed folds: re-apply
          'foldlevel', then do "zv": View cursor line.
          Also forces recomputing folds.  This is useful when using
          'foldexpr' and the buffer is changed in a way that results in
          folds not to be updated properly.