fix(nvim): beware of cleared cache
I just ran into this:

    $USER/.config/nvim/lua/wincent/foldexpr.lua:83: attempt to get length of local 'old_markers' (a nil value)
    stack traceback:
            $USER/.config/nvim/lua/wincent/foldexpr.lua:83: in function </Users/greg.hurrell/.config/nvim/lua/wincent/foldexpr.lua:75>

Somehow, `cache[bufnr]` is returning `nil`.

Maybe the `on_reload` callback is firing and blowing it away? That
sounds like the most likely explanation, based on `:h
nvim_buf_attach()`:

    • on_reload: Lua callback invoked on reload. The entire
      buffer content should be considered changed. Args:
      • the string "reload"
      • buffer handle

Whatever the case, add an `if`/`else` to make sure we don't try to count
the length of a `nil` value.