refactor(nvim): turn `:OpenOnGitHub` into a wrapper for `:GBrowse`
This _does_ suck a bit, in the sense that if I use it with a selection
in a buffer that has unpushed changes, it will blow up:

    Error executing Lua callback: vim/_editor.lua:0: nvim_exec2():
    Vim(echoerr):fugitive: can't browse to unpushed change

I do that often (and the fact that the line numbers don't match doesn't
bother me at all; I'm usually just trying to view the file in the remote
state and I don't care). Oh well. The workaround of doing it _without_ a
selection seems to work.

Note that I am checking to see if vim-fugitive and vim-rhubarb are
loaded by using this code:

    if plugin == candidate:sub(-#plugin) then
      return true
    end

ie. if we take the last `#plugin` characters of a full path in the
`'runtimepath'`, and the result is equal to `plugin`, then we have a
match.

I would have written `string.match(candidate, pattern)`, but you have to
escape stuff (like the `-` in `vim-fugitive`) and I don't want to use an
abomination like this:

    local function escape_pattern(str)
      -- In Lua patterns, a bunch of stuff is special.
      return str:gsub('[%(%)%.%%%+%-%*%?%[%]%^%$]', '%%%1')
    end

in order to do so...