chore(dotfiles): add a couple more .luarc.json files
To suppress warnings about undefined globals.

Note: this only works if you open Neovim with a path in one of the
subdirectories; if you open something else, the project root will be
determined to be the repo root, and opening one of the files in the
subdirectories doesn't change that. There might be a way to make that
work better, but I'll leave that challenge for another day.