perf(nvim): make opening large Lua files actually feasible
Motivation: I have a 48 megabyte Lua file (don't "at" me, bro) in the
Command-T project:

- https://github.com/wincent/command-t/blob/main/data/wincent/commandt/benchmark/configs/matcher.lua

```
$ wc data/wincent/commandt/benchmark/configs/matcher.lua
  646902  647160 50555040 data/wincent/commandt/benchmark/configs/matcher.lua

$ ls -lh data/wincent/commandt/benchmark/configs/matcher.lua
-rw-r--r--@ 1 wincent  staff    48M 20 Jan 11:50 data/wincent/commandt/benchmark/configs/matcher.lua
```

ie. more than 646,000 lines, and it was painfully slow to open and
painfully slow to do anything in it in Neovim; even moving the cursor
was slow, and mutating the buffer even more so. Vim, on the other hand,
is lightning fast, both to open, but also to scroll and edit.

Summarizing my investigation:

- I found that `nvim -u NONE` was fast.
- Disabling all plugins did _not_ make it fast.
- `nvim --startuptime` showed lots of time being spent in
  `vim.treesitter.highlight`, apparently.
- Configuring treesitter to not use highlighting (or any other modules,
  such as "indent", and so on) did not help.
- It turns out that even if you turn off all treesitter modules, the
  Neovim runtime files unconditionally call `vim.treesitter.start()`,
  which has a cost.
- Likewise, although you turn off the slow `nvim_treesitter#indent()`
  `'indentexpr'`, it just causes the system to fall back to the
  also-slow `GetLuaIndent()`.
- My `'foldexpr'` isn't a problem for editing and scrolling, but it
  _might_ be slowing down initial load (`--startuptime` shows lots of
  time spent reading the view file under `~/.config/nvim/view/`).

Solution:

1. Turn off treesitter modules for large files.
2. Additionally call `vim.treesitter.stop()` for large Lua files; I
   don't know whether a similar fix will be needed for other filetypes,
   but if so, I will apply it as I discover the need.
3. Turn off syntax too.

With this change, initial load is still slow, but scrolling and editing
is fast , and that is a good enough improvement for me to warrant
committing this.

FWIW, I did try adding `vim.opt_local.viewoptions:remove('folds')` to
my large-buffer handling here, and it made no difference (likewise for
the other settings under `'viewoptions'`). It seems that something else
is responsible for the time being spent opening the file. Right now
`--startuptime` shows these two main painpoints:

First, we observe a big delay (about 1.2s) when the runtime file
unconditionally runs `vim.treesitter.start()`:

    075.138  001.707: reading ShaDa
    275.914  000.051  000.051: sourcing $HOME/.config/nvim/ftplugin/lua.lua
    277.492  000.033  000.033: sourcing /opt/homebrew/Cellar/neovim/0.10.3/share/nvim/runtime/ftplugin/lua.vim
    1535.676  000.256  000.256: require('vim.treesitter.highlighter')
    1535.769  1258.236  1257.981: sourcing /opt/homebrew/Cellar/neovim/0.10.3/share/nvim/runtime/ftplugin/lua.lua
    1536.736  000.065  000.065: require('wincent.vim.is_large_buffer')
    1538.244  000.339  000.339: sourcing /opt/homebrew/Cellar/neovim/0.10.3/share/nvim/runtime/syntax/lua.vim

Then we see another big delay (over 900ms) when processing the view
file:

    1556.233  000.060  000.060: require('editorconfig')
    1556.442  000.011  000.011: sourcing $HOME/.config/nvim/view/~=+code=+wincent=+aspects=+nvim=+files=+.config=+nvim=+pack=+bundle=+opt=+command-t=+data=+wincent=+commandt=+benchmark=+configs=+matcher.lua=
    2483.707  1143.193: opening buffers