fix(violentmonkey): try harder to apply Fastmail patch
Noticed userscript wasn't working; unclear whether Orion's fault, or
Violentmonkey's. All manner of reloading and refreshing and restarting
didn't help. Later, saw in tooltip a warning that the script matched but
didn't match when the page first loaded, and that I should try
`@match`-ing the entire site. So, that's what I am doing here (note the
script already has a `location` check internally, so this is ok), and it
seems to work again.