chore: specify branch for nvim-clipper
So that `update-bundle` knows what to do.

It defaults to "master" if unset.

When can we flip the default? Not yet. This prints out the branch for
each submodule:

    git submodule foreach 'echo "$(pwd) -> $(git rev-parse --abbrev-ref HEAD)"'

Right now, we have 33 submodules with "main" as their branch, and 35
with "master". Then we have 4 odd ones out:

- shellbot: wincent
- source-code-pro: release
- git-cipher: next
- tinted-theming/schemes: spec-0.11

I was curious to see when this all began, so I went back and checked.
It's been ~~84 years~~ 5 years, since the movement from `master` to
`main` started:

- https://github.com/github/renaming
- https://about.gitlab.com/blog/2021/03/10/new-git-default-branch-name/
- https://sfconservancy.org/news/2020/jun/23/gitbranchname/