fix(dotfiles): work around garbage in `.tmux` script
I was seeing a garbage string (looking like an escape sequence, and
maybe a hexadecimal hash - not sure of the source of it) inserted into
the right pane every time I used this script.

A few changes in here, but only the last one matters; the first two were
made while troubleshooting:

1. Run `vim` last, not first, because it seemed to be interfering with
   the `git` command (ie. I could comment it out, and the problem would
   go away).
2. Don't use the pager (again, trying to rule that out as a factor);
   this turned out not to fix the problem but I prefer it this way
   anyway so that I can start typing immediately.
3. Wait for things to settle (0.5s) before actually attaching. This is
   the only thing that works. 0.1s does not fix the problem, so I tried
   to pick a more generous value.

No idea what is causing this. Is it a Neovim thing? A Command-T thing? A
Sequoia thing? A kitty thing? A tmux thing? All of these things change
regularly on my system or have changed recently.

Additional notes:

- Moving the sleep inside the shell doesn't help (which is
  particularly mind-bending, IMO, because you can see the right pane
  getting corrupted irrespective of the delay):

      tmux send-keys -t dot:vim.left "sleep 2; vim -c CommandT" Enter

- Taking Command-T out of the equation doesn't help either:

      tmux send-keys -t dot:vim.left "sleep 2; vim" Enter

- Removing the final `attach` call also makes the problem go away.

- Switching to Vim (ie. `\vim` instead of my `vim` → `nvim` alias) also
  makes the problem go away.