feat(dotfiles): improve tmux copy bindings
Make them a bit more (but not entirely) Vim-like. Inspired by:

- https://github.com/wincent/wincent/issues/154

`y` is currently unbound, so can be trivially set to the same as
`Enter`.

`v`, `C-v`, and `V` all have existing bindings. We make them a bit more
Vim-like:

- `v`: Before, toggled rectangular selection mode. Now, the first `v`
  engages `VISUAL` mode and the second `v` disengages it.
- `V`: Before, always started a new `VISUAL LINE` selection. Now, the
  first `V` engages `VISUAL LINE` mode and the second `V` disengages it.
- `C-v`: Before, toggled rectangular selection mode. Now, the first
  `C-v` engages `VISUAL BLOCK` mode and the second `C-v` disengages it.

This is different from Vim, where if I press `v`, `V`, then `C-v`, I'll
first enter `VISUAL` mode, and then `VISUAL LINE` and then `VISUAL
BLOCK`.

I could make this more like Vim with a bit of work: ie. I could set a
variable when entering each mode to show which mode I'm in (eg. `VISUAL`
could be `1`, `VISUAL LINE` could be `2`, and `VISUAL BLOCK` could be
`3`), and then I could check the value in each of the bindings; ie:

- If you hit `v`, I could set the value to `1`.
- If you then hit `V`, I go to `VISUAL LINE` if the value is not `2`;
  otherwise I would clear the selection.

Not sure if I can be bothered doing that right now (I'd have to change
all the places that you can get into the `VISUAL` modes, such as mouse
bindings), but I think I'll do it in a future iteration.