feat(nvim): add LSP progress display
I've had this uncommitted for a while now, because I was thinking about
doing fancy stuff with it, like a progress indicator. But in practice
all I want is the ability to see that something is happening in the LSP,
and this already does that. The code could be cleaned up, but I'll do
that later. Maybe.