refactor(nvim): use `vim.opt_local` instead of `setlocal` in plaintext
As noted in the comment of my `setlocal()` implementation:

```
-- This was originally a workaround due to buggy behavior of `vim.opt_local`:
--
--    https://github.com/neovim/neovim/issues/14670
--
-- That one is apparently resolved now, but I am leaving it as a convenience
-- wrapper in any case. Its job is to match the magical behavior of Vim's
-- `:setlocal` command, which can apply window-local or buffer-local settings,
-- even for settings like `'spell'` which are labeled as "window"-level settings
-- in the docs, but which actually need to be buffer-level ones in order to
-- provide a good user experience. `vim.api.nvim_buf_set_option()` errors if you
-- try to set `'spell'`, but `vim.api.nvim_set_option_value()` does not, and
-- it ends up setting it at a buffer level, not a window level, so we use that
-- here.
```

I still have a bunch of places using it, but to start with a small step
towards moving to `vim.opt_local`, let's stop using it in this one
place (the implementation of our `plaintext()` function). In testing,
seems to work fine. Settings appear to be buffer-sticky and not
window-sticky, which is what we want.

Before going further with this (ie. before replacing more calls to
`setlocal()` with usage of `vim.opt_local`) I need to test to make sure
that the behavior is as desired.