fix(dotfiles): avoid tmux "Not in a mode" spam
This binding was causing the scroll-up mouse action in Vim to cause tmux
to spam "Not in a mode" in the status area at the bottom of the screen.

TBH, I am not sure if "pane_in_mode" will ever be true in the root
bindings table, so I might be able to simplify further from here.