refactor(nvim): rename file
Because the inconsistency with the other file and the plug-in name
itself is bugging me.

- Other file: aspects/nvim/files/.config/nvim/after/plugin/indent-blankline.nvim.lua
- Plug-in: aspects/nvim/files/.config/nvim/pack/bundle/opt/indent-blankline.nvim

ie. they both use a hyphen instead of underscore

I can't rename this file to use a hyphen because it relies on
autoloading (ie. via `wincent.plugins.indent_blankline`). Well, I guess
I could do `wincent.plugins['indent-blankline']`, but that seems rather
ugly.

Instead, rename to `ibl.lua`, which is consistent with what you do to
access the plug-in programmatically (ie. `require('ibl')`).