feat(violentmonkey): add script to autoclose the Zoom interstitial
Note that I discovered a sharp edge here in the `templates()` function.
Initially, I tried to write the task that creates the `index.html` with:

    task('create index.html', async () => {
      await template({
        path: '~/Sites/UserScripts/index.html',
        src: resource.template('UserScripts/index.html'),
      });
    });

but found it was complaining that there was already a directory at
`path`... Turns out that was created by:

    task('fill templates', when('wincent'), async () => {
      for (const src of resource.templates('UserScripts/*/*.js')) {
        await template({
          path: path('~/Sites/UserScripts').join(...src.last(2)),
          src,
        });
      }
    });

which runs first. The `resource.templates()` glob was looking at
everything in `Userscripts/` and seeing the file `index.html`, and
creating a directory with that name, it seems.

"Fixed" this by storing the template one level higher:

    task('create index.html', async () => {
      await template({
        path: '~/Sites/UserScripts/index.html',
        src: resource.template('index.html'),
      });
    });

but in the future I should implement a real fix.