refactor: prepare to run directly with Node.js (ie. no build)
Adding two new options to pave the way for taking advantage of
type-stripping added to Node in v22.6.0 (note: I don't even know if I'm
going to do this, but I want to be prepared):

- https://nodejs.org/api/typescript.html#type-stripping

Changes to source needed for these errors:

```
fig/readAspect.ts:6:9 - error TS1484: 'Aspect' is a type and must be imported using a type-only import when 'verbatimModuleSyntax' is enabled.

6 import {Aspect, assertAspect} from './types/Aspect.js';
          ~~~~~~

fig/readProject.ts:6:9 - error TS1484: 'Project' is a type and must be imported using a type-only import when 'verbatimModuleSyntax' is enabled.

6 import {Project, assertProject} from './types/Project.js';
          ~~~~~~~

Found 2 errors in 2 files.

Errors  Files
     1  fig/readAspect.ts:6
     1  fig/readProject.ts:6
```