feat(neovim): add `<leader>ppp` binding
So, now we have:

- `<leader>p` print file name
- `<leader>pp` print file name and copy it to the clipboard
- `<leader>ppp` print absolute path and copy it to the clipboard

Built-in `CTRL-G` (see `:help CTRL-G`) already does this in part, but it
is unpredictable. If the current file is inside the present working
directory, it prints a relative path. If it is outside, it prints an
absolute path. You can force printing the absolute path even in the
first scenario by doing `{count}CTRL-G`, but the variability is
annoying.

Note that normally sharing prefixes across bindings is a no-go because
they produce a delay while Neovim waits to disambiguate between "user
typed `<leader>p` and then a bit later presses `p` again, meaning I
should use the `<leader>pp` mapping" and "user typed `<leader>p` and
then nothing else, so I should use the `<leader>p` mapping". In this
case, however, it's fine, as these mappings aren't used in
"performance-critical sections" of the editing workflow.