feat(violentmonkey): add a script for suppressing Atlassian AI
Not all the AI features, but the most annoying one (the thing that sits
above all search results, repeatedly shifting the layout of everything
underneath it so that you can't actually click on the results).

Hopefully this obfuscated class name won't change too often, but it has
been stable for (at least) the last week or so, so fingers crossed.