fix(nvim): try to avoid spurious "duplicate-doc-field" diagnostics
I was hopefully that this would fix the spurious diagnostics seen
in plug-in files, but it doesn't. Nevertheless, seems like a
worthwhile optimization to implement (as it will avoid potential
double-traversals), even if the comment is perhaps a bit misleading...
Going to leave the comment as-is for now and will alter if subsequent
investigation sheds any light on the issue.

FWIW, here's an example of what I'm trying to fix: in
`~/.config/nvim/pack/bundle/opt/nvim-cmp/lua/cmp/types/cmp.lua` the
whole buffer is lit up like a Christmas tree, like you see in this
excerpt:

        3 ---@class cmp.ContextOption
    ⚐   2 ---@field public reason cmp.ContextReason|nil
                           └──── duplicate-doc-field: Duplicate defined fields `reason`.
        1
      45  ---@class cmp.ConfirmOption
    ⚐   1 ---@field public behavior cmp.ConfirmBehavior
                           └──── duplicate-doc-field: Duplicate defined fields `behavior`.
    ⚐   2 ---@field public commit_character? string
                           └──── duplicate-doc-field: Duplicate defined fields `commit_character`.

Same for any Lua plug-in with annotations.