docs(nvim): note where I should be using `LspAttach`
Not suffering any problems right now because none of the LSP configs I
use supply an `on_attach` (which would overwrite my own). But adding
this note here that the "right" way to respond to an attach event when
using nvim-lspconfig is to do it via an autocmd instead. Context:

- https://github.com/neovim/nvim-lspconfig/issues/3827
- https://github.com/neovim/neovim/issues/33577#issuecomment-2822066227