fix(nvim): definitively(?) fix "duplicate-doc-field" diagnostics
This seems to be the trick to suppressing these. My guess is that by
passing plug-in files into "workspace.library", lua-language-server
considers that you only want to know what types are in there, for the
purposes of editing other files; if you try to edit one of those files
directly, then it loads the content once (because you're editing it) and
again (because you told the language server to take note of its contents
as a library file).

There may be some more wrinkles to iron out here, but this does at least
make the diagnostics go away.