fix(nvim): make workspace overrides work
This is quite subtle.

If the language server detects a workspace, and that workspace has a
".luarc.json" or ".luarc.jsonc", defer to it (ie. bail with an early
`return` and apply no additional settings).

When applying defaults, I was going to do something "clever" here and
have it only take place if I detected you were working somewhere under
"~/.config/nvim". Turns out that is annoying to do in my case. My
dotfiles live under "~/code/wincent", and "~/.config/nvim" is actually a
symlink to a file somewhere else; so, if you check to see whether a path like
"~/code/wincent/aspects/nvim/files/.config/nvim/pack/bundle/opt/nvim-cmp/lua/cmp/types/cmp.lua"
is under "~/.config/nvim" via simple prefix comparison, your detection
will fail. I may still fix this subsequently, by turning both into
"real" paths (dereferencing symlinks), but for now, I just apply the
defaults unconditionally when doing the fallback.

Finally, note the changes I made to `root_markers`: this is the subtle
part that I was referring to above. `root_markers` are currently set to:

    root_markers = {
      '.luarc.json',
      '.luarc.jsonc',
      '.luacheckrc',
      '.stylua.toml',
      'stylua.toml',
      'selene.toml',
      'selene.yml',
      '.git',
    },

- https://github.com/neovim/nvim-lspconfig/blob/1cb30b1bafe5a63a5c6ac20dc39f83487df38855/lsp/lua_ls.lua#L73-L81

I'm on the latest Neovim release (v0.11.2), and the way that uses
`root_markers` is a bit surprising:

    for _, marker in ipairs(opts._root_markers) do
      local root = vim.fs.root(bufnr, marker)
      if root ~= nil then
        config.root_dir = root
        break
      end
    end

- https://github.com/neovim/neovim/blob/a73904168a09b04b00dc5a4f11c48b10a9e43d4a/runtime/lua/vim/lsp.lua#L706-L707

That is, it does a _depth-first_ search, looking first for
".luarc.json", then ".luarc.jsonc" etc, and it will bail on the first
one that it finds. This produces a very confusing result for a plug-in
like Command-T, which is deeply nested inside my dotfiles. Instead of
finding Command-T's own ".luarc.jsonc", it first looks for ".luarc.json"
and finds the one many layers higher up in the filesystem at the root of
my dotfiles repo, and eagerly decides that's the best possible match.

The fix is in the docs, which show you can pass nested lists like I am
in this commit. What that means is it will search for _any_ of
".luarc.json", ".luarc.jsonc" etc, walking upwards, and then fall back
to the next nested list (or string), looking for that (in this case,
".git").

The current master branch looks a bit different:

    if not config.root_dir and opts._root_markers then
      validate('root_markers', opts._root_markers, 'table')
      config = vim.deepcopy(config)

      config.root_dir = vim.fs.root(bufnr, opts._root_markers)
    end

- https://github.com/neovim/neovim/blob/807bc00dd68dccb55d5aa12e0a568d2551b08d04/runtime/lua/vim/lsp.lua#L730

So I don't know if nested marker lists will continue to work, but
looking at the commit where the change was made, looks like the intent
is that `fs.root()` will itself be able to handle nested lists:

- https://github.com/neovim/neovim/commit/2f0fbdaa480c5b014b68bfdb85aa105d0bef3637

In any case, if it breaks, we can define a `root_dir()` callback and
handle it on our own if we ever need to. Will cross that bridge if and
when we come to it.

Tested this in a few places:

- My dotfiles (uses dotfiles repo as root, because it has a
  ".luarc.json").
- Command-T plugin (uses its own directory as root, because it has a ".luarc.jsonc").
- nvim-cmp (uses its own directory as root, because it has a ".luacheckrc")
- In my Hammerspoon files (uses its own directory as root, because it
  has a ".luarc.json")
- In a random scratch directory (doesn't find a workspace, because there
  are no markers; ends up using fallback values).