refactor(nvim): decide whether to use fallback values based on heuristic
TBH, I'm not even really sure I'll hit this path that often, but this
still seems the "technically correct" way to do it.

In the common case, we'll actually see the top-level ".luarc.json" and
use that instead (which is why I've added the necessary settings to that
one in this commit).