feat(dotfiles): add `--engine=auto` to `~/.rgrc` and Ferret settings
So I can use PCRE2 lookaround when needed. Note that for most `rg` use
cases, the positive forms aren't that useful. That is, if you want "foo
followed by bar", it's often good enough to just search for "foobar".
But the negative forms are great when you want to narrow down to a
subset of the results; eg. if you want to search for a naked identifier
but not the quoted equivalent (ie. `foo...` but not `"foo...`) you can
do something like `(?<!")foo`.

Note that adding the setting to `~/.rgrc` alone is not enough for my
most common use case (using Ferret in Neovim), because Ferret passes
`--no-config` by default. And note that I don't want to add
`--engine=auto` as a default to Ferret itself, because as `man rg`
notes:

> Note that the pcre2 engine is an optional ripgrep feature. If PCRE2
> wasn't included in your build of ripgrep, then using this flag will
> result in ripgrep printing an error message and exiting.

I haven't tested whether this happens just for `--engine=pcre2` or
also for `--engine=auto` (because my build of `rg` _does_ have PCRE2
support), so I don't want to risk breaking things for anybody. Also, I
don't want to break things for anybody using an old version of `rg`,
which might not have the `--engine` option.

As such, I'm tweaking my local settings to pass
`--engine=auto` through to Ferret. You might ask, seeing as
`~/.rgrc` has the same overrides/extensions in it as I'm adding in
`~/.config/nvim/after/plugin/ferret.lua`, why don't I just drop the
`--no-config` setting and let Ripgrep read the config file? It's
basically because I want to maintain a layer of separation, so that
changes to the config don't affect Neovim unless I have vetted those
changes as being compatible with it. Maybe this is overly cautious, but
it is what it is, and I like the explicitness (I can just look in the
Neovim config and know exactly how `rg` is going to be called).