refactor(nvim): prefer function form of `vim.cmd()`
Noticed that `--startuptime` log contains a bunch of lines like this:

    020.484  000.026  000.014: sourcing nvim_exec2() called at ~/code/wincent/aspects/nvim/files/.config/nvim/init.lua:0

corresponding to `vim.cmd('...')` calls. Seems that switching to the
function form avoids this.

Note that there can be subtle differences between user commands and
built-ins:

- This works: `vim.cmd.cabbrev('sord sort')`
- But this doesn't: `vim.cmd.Abolish('aboud about') (it prints `E474:
  Invalid argument (not a keyword: 'carp crap')`)
- Instead, I had to write: `vim.cmd.Abolish({ 'aboud', 'about' })`

Not sure if that's an implementation detail of `:Abolish` (source code
is inscrutable to me), or something intrinsic about built-ins vs user
commands.