refactor(tmux): use hooks instead of passing `-H` in `copy-mode` bindings
Implements the suggestion from Mr Meowski on the linked issue.

This ends up removing all the `-H` switches that were added in
bf50fd697538289c ("feat(dotfiles): improve copy mode navigation in
tmux", 2024-08-23) in favor of turning off the default indicator using a
hook.

This is better because it means I don't have to redefine the default
bindings (unless there is good reason to: for bindings where this is the
case I have left them in the config), and if any new default bindings
are ever added I will automatically handle those too.

Closes: https://github.com/wincent/wincent/issues/155