refactor: switch to tinted-terminal from tinted-kitty
tinted-kitty repo got archived:

- https://github.com/tinted-theming/tinted-kitty/commit/de6f888497f2c6b2279361bfc790f164bfd0f3fa

and contents folded into tinted-terminal repo instead, along with
configs for a bunch of other terminals:

- https://github.com/tinted-theming/tinted-terminal/tree/main/templates

So, given that the config now includes a bunch of keys for other
terminals, I had to teach the `support/tinted-builder.rb` script to take
an optional key name to filter on.

Also note that they changed the directory name where output files get
written: it was "colors/" but is now "themes/". I could have taught the
builder to write to a different directory, but instead I am just letting
it do its thing and I instead changed my own config to look in the new
location.

Ideally this would be the end of the breaking changes, but given the
history of the project I am not hopeful.

* aspects/nvim/files/.config/nvim/pack/bundle/opt/base16-nvim eb101e5...f4d7848 (1):
  > chore: run theme update from super project

* vendor/tinted-theming/schemes 5a775c6...c279b1e (1):
  > Adds base16 everforest-dark-medium theme variant (#64)

* vendor/tinted-theming/tinted-kitty de6f888...0000000:

* vendor/tinted-theming/tinted-terminal 0000000...99053fb0 (50):
  > doc: AppleScript hack to auto-refresh Ghostty shell (#22)

* vendor/tinted-theming/tinted-vim ea3cbd6c...da53d414 (5):
  > Update with the latest tinted-theming colorschemes
  > Update with the latest tinted-theming colorschemes
  > Check both(?) legacy properties (#108)
  > Update with the latest tinted-theming colorschemes
  > Add options to enable/disable bold,italic,strikethrough,underline,und… (#102)