feat: silence Yarn console spam
Since 1e90acabd58e ("refactor: run Yarn v4 using corepack, instead of
vendored v1", 2025-06-10), runs have included this Yarn console spam on
every run:

    $ ./install dotfiles
    [info]    Installing TypeScript
    ➤ YN0000: · Yarn 4.9.2
    ➤ YN0000: ┌ Resolution step
    ➤ YN0000: └ Completed
    ➤ YN0000: ┌ Fetch step
    ➤ YN0000: └ Completed
    ➤ YN0000: ┌ Link step
    ➤ YN0000: └ Completed
    ➤ YN0000: · Done
    [info]    Cleaning
    [info]    Generating TypeScript types
    [info]    Checking TypeScript source
    [info]    Running Fig
    [info]    Working from root: ~/code/wincent
    [info]    Running tests
    ...

even though we're calling yarn with the `--silent` flag. This code
comment suggests that `--silent` is a no-op in v4's `run` command, at
least:

- https://github.com/yarnpkg/berry/blob/267725122fef2961c46ad1b6c970f191b4c222e1/packages/plugin-essentials/sources/commands/run.ts#L59-L64

and this one:

- https://github.com/yarnpkg/berry/blob/267725122fef2961c46ad1b6c970f191b4c222e1/packages/plugin-essentials/sources/commands/install.ts#L111-L119

claims that using it should actually produce a warning in the case of
`install`, which doesn't seem to be the case.

So, I'm removing it from both `run` and `install` invocations this
commit.

In any case, there is a `logFilters` functionality that should allow you
to suppress the unwanted messages:

- https://yarnpkg.com/configuration/yarnrc#logFilters

by inserting something like this in your `.yarnrc.yml`:

```
logFilters:
  - code: YN0000
  - level: discard
```

However, as noted here:

- https://github.com/yarnpkg/berry/issues/6221

there is an `if` that lets "unnamed" messages through unconditionally,
regardless of log filter configuration:

- https://github.com/yarnpkg/berry/blob/db6210f48355d2986e965f90009b22f18d3b6342/packages/yarnpkg-core/sources/formatUtils.ts#L535-L537

This appears to be intentional and not an oversight:

- https://github.com/yarnpkg/berry/pull/1907#discussion_r502291653

So, even using a more aggressive filter, like:

```
logFilters:
  - pattern: '*'
  - level: discard
```

will have no effect.

Looks like the only way to get what we want then, is to redirect to good
ol' `/dev/null`.

Now, runs again start like this:

    $ ./install dotfiles
    [info]    Installing TypeScript
    [info]    Cleaning
    [info]    Generating TypeScript types
    [info]    Checking TypeScript source
    [info]    Running Fig
    [info]    Working from root: ~/code/wincent
    [info]    Running tests
    ...