refactor(nvim): use `once()` helper
I don't like the subtlety of having to use a global variable to ensure
we run this callback once only (subtle, because idiomatically, you would
reach for a module-local variable, but then you realize that `ftplugin/`
files aren't like normal Lua modules because they get re-evaluated every
time the filetype is detected). I can see myself thoughtlessly
refactoring this in a way that breaks in the future.

Sure, I was stashing the `{ registered = true }` sentinel in the global,
so that is at least self-documenting, but it doesn't explain _why_ it
has to be stashed in a global instead of just a module-local variable.

Abstracting that away inside a `once()` helper makes this much harder to
accidentally break, and as a bonus, it _looks_ a bit more "functional"
and less "procedural", although the reality is that both are the same.