fix(nvim): don't blow up when opening `.gitignore`
As noted in the issue, this seems a new thing, so not sure why it
cropped up. The tree-sitter parser at play here is called "gitignore",
and that's the standard filetype too, but I was using "ignore" as a
filetype for a number of different tools that all follow gitignore-ish
rules.

    Error detected while processing BufReadPost Autocommands for ".gitignore"..FileType Autocommands for "ignore":
    Error executing lua callback: .../neovim/0.11.3/share/nvim/runtime/lua/vim/treesitter.lua:431: Parser could not be created for buffer 1 and language "ignore"

The other thing I tried was something like:

    vim.treesitter.language.register(parser, { filetypes[parser] })

in `lua/wincent/treesitter/setup.lua`; ie. that would do:

    vim.treesitter.language.register('gitignore', { 'ignore' })

That does indeed make the error go away, but it breaks all the other
filetypes with special parsers. For example, `:Inspect` no longer worked
in SSH config files, Markdown files etc.

Seems the easiest fix, then, is to just make all these ignore files use
"gitignore" as the filetype.

Closes: https://github.com/wincent/wincent/issues/159